+++ src/common_includes/can_id_list.h
@@ -1,0 +1,6 @@
+#ifndef CAN_ID_LIST_H
+#define CAN_ID_LIST_H
+
+#define CAN_ID_COMMAND   (0x111U)
+
+#endif
