+++ src/common_includes/logging.c
