+++ src/receiver.c
@@ -1,5 +1,5 @@
-#include "can_socket.h"
-#include "logging.h"
+#include "common_includes/can_socket.h"
+#include "common_includes/logging.h"
 #include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
